# Contracts/Request/UserLoginDTO.cs
using System.ComponentModel.DataAnnotations;

namespace TVShowTracker.Contracts.Request
{
    public class UserLoginDTO
    {
        [Required]
        [EmailAddress]
        public string Email { get; set; }

        [Required]
        public string Password { get; set; }
    }
}