# Contracts/Request/UserRegistrationDTO.cs
using System.ComponentModel.DataAnnotations;

namespace TVShowTracker.Contracts.Request
{
    public class UserRegistrationDTO
    {
        [Required]
        [EmailAddress]
        public string Email { get; set; }

        [Required]
        public string Password { get; set; }

        [Required]
        public string Username { get; set; }
    }
}